+++ generator/codegen/testdata/test-redacted.gen.go
@@ -1,6 +1,0 @@
-/*
- * Copyright (C) 2020-2024 Arm Limited or its affiliates and Contributors. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package client